backend/requirements.txt
Flask
flask-cors
flask-jwt-extended
python-dotenv
PyMySQL
requests